extract listTypes array to be references when needed
